Pass key-value tags as objects
Implement passing key-value tags as objects. This reduces the chance of errors, and also allows implementing sanitation of tags.

The old format is still supported, and this is relying on the existing
parsing and formatting functionality. If the array format would be
removed, some of the parsing would become redundant and tags could be
passed as an object until they are serialised as a string over the wire.

Fixes #55.