remove localhost fallback for host option
According to dgram socket, if the address provided is a hostname such as
localhost, it will query DNS to resolve it. dns.lookup() has a fast path
for IP addresses, when omitting a value for host.

Also, from
https://nodejs.org/api/dgram.html#dgram_socket_send_msg_offset_length_port_address_callback:

"If address is not provided or otherwise falsy, '127.0.0.1' (for udp4
sockets) or '::1' (for udp6 sockets) will be used by default."